+++ src/WPF/MsgBox/Native.cs
@@ -1,0 +1,15 @@
+﻿
+using System;
+using System.Runtime.InteropServices;
+
+namespace MsgBoxEx
+{
+    class Native
+    {
+        [DllImport("user32.dll")]
+        public static extern IntPtr GetForegroundWindow();
+
+        [DllImport("user32.dll")]
+        public static extern uint GetWindowThreadProcessId(IntPtr hWnd, out uint lpdwProcessId);
+    }
+}