+++ docs/Changelog.md
@@ -3,7 +3,8 @@
 
 - All PowerShell dependencies except the scipt host were removed
   in order to achieve first class Windows 7 compatibility!
-  
+- Fix message box not working when ontop is enabled.
+
 
 5.4.9.7 Beta (2021-08-28)